requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
apipkg==1.4               # via execnet
appnope==0.1.0            # via ipython
attrs==17.4.0             # via pytest
backcall==0.1.0           # via ipython
cached-property==1.4.2    # via docker-compose
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
coverage==4.5.1
decorator==4.3.0          # via ipython, traitlets
django-choices==1.6.0
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==2.0.7
django-rest-swagger==2.1.2
django==2.0.4
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
docker-compose==1.21.0
docker-pycreds==0.2.2     # via docker
docker==3.2.1             # via docker-compose
dockerpty==0.4.1          # via docker-compose
docopt==0.6.2             # via docker-compose
execnet==1.5.0            # via pytest-xdist
gevent==1.2.2
greenlet==0.4.13          # via gevent
gunicorn==19.7.1
idna==2.6                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.3.1
itypes==1.1.0             # via coreapi
jedi==0.12.0              # via ipython
jinja2==2.10              # via coreschema
jsonschema==2.6.0         # via docker-compose
markupsafe==1.0           # via jinja2
model-mommy==1.5.1
more-itertools==4.1.0     # via pytest
openapi-codec==1.3.2      # via django-rest-swagger
parso==0.2.0              # via jedi
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython
psycopg2-binary==2.7.4
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest
pygments==2.2.0           # via ipython
pyjwt==1.6.1              # via djangorestframework-jwt
pytest-django==3.2.1
pytest-forked==0.2        # via pytest-xdist
pytest-xdist==1.22.2
pytest==3.5.0             # via pytest-django, pytest-forked, pytest-xdist
pytz==2018.4              # via django
pyyaml==3.12              # via docker-compose
requests==2.18.4
simplegeneric==0.8.1      # via ipython
simplejson==3.13.2        # via django-rest-swagger
six==1.11.0               # via django-environ, django-extensions, docker, docker-compose, docker-pycreds, dockerpty, model-mommy, more-itertools, prompt-toolkit, pytest, traitlets, websocket-client
texttable==0.9.1          # via docker-compose
traitlets==4.3.2          # via ipython
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
websocket-client==0.47.0  # via docker, docker-compose
